+++ rust/cubestore/src/metastore/partition.rs
@@ -6,10 +6,8 @@ use crate::metastore::{IdRow, MetaStoreEvent};
 use crate::rocks_table_impl;
 use crate::table::Row;
 use byteorder::{BigEndian, WriteBytesExt};
-use chrono::Utc;
 use rocksdb::DB;
 use serde::{Deserialize, Deserializer};
-use std::ops::Sub;
 
 impl Partition {
     pub fn new(index_id: u64, min_value: Option<Row>, max_value: Option<Row>) -> Partition {
@@ -76,12 +74,6 @@ impl Partition {
         p
     }
 
-    pub fn update_last_used(&self) -> Self {
-        let mut new = self.clone();
-        new.last_used = Some(Utc::now());
-        new
-    }
-
     pub fn get_index_id(&self) -> u64 {
         self.index_id
     }
@@ -100,12 +92,6 @@ impl Partition {
 
     pub fn main_table_row_count(&self) -> u64 {
         self.main_table_row_count
-    }
-
-    pub fn is_used(&self, timeout: u64) -> bool {
-        self.last_used
-            .map(|time| Utc::now().sub(time.clone()).num_seconds() < timeout as i64)
-            .unwrap_or(false)
     }
 }